Bunch of poorly thought out casting to get things happy

+++ src/whisper/file/file.rs
@@ -55,7 +55,7 @@ impl<'a> WhisperFile<'a> {
     pub fn calculate_write_ops(&self, current_time: u32, point: point::Point) -> Vec<WriteOp> {
         // let current_time = get_time().sec as u32;
         let hp_ai_option = self.header.archive_infos.iter().find(|ai|
-            (current_time - point.timestamp) < ai.retention
+            (current_time - point.timestamp) < ai.retention as u32
         );
 
         let write_ops = vec![];
